rework api documentation
