+++ api/interactivity.md
@@ -1,16 +1,16 @@
-<a name="api-sg-functions-tooltip"></a>
-## Tooltip Functions ##
+## Adding interactive features ##
 
-<a name="api-sg-addTooltipToPoints"></a>
-**`SimpleGraph.addTooltipToPoints(tooltipFunction[, options])`** : Add tooltip function to the points on the graph. Does not add tooltips to the lines connecting points, if they were added. That is handled by addTooltipToLines(). You can check the series name in the callback's returned SVG element or the class to determine type, regular lines are `.sg-plotted-line` and lines drawn from connecting points are `.sg-line`.
+<a name="addTooltipToPoints">#</a> *sg*.**addTooltipToPoints**(*tooltipFunction*[, *options*])
 
-<table>
+Add tooltip function to the points on the graph. Does not add tooltips to the lines connecting points, if they were added. That is handled by addTooltipToLines(). You can check the series name in the callback's returned SVG element or the class to determine type, regular lines are `.sg-plotted-line` and lines drawn from connecting points are `.sg-line`.
+
+<table style="font-size:0.9em;">
   <tbody>
     <tr>
       <th>Name</th><th>Type</th><th>Description</th>
     </tr>
     <tr>
-      <td>tooltipFunction</td><td><a href="./defs.md#api-sg-defs-tooltip">tooltipTextFunction</a></td><td>Callback function that handles the dynamic text appearing in the tooltip.</td>
+      <td>tooltipFunction</td><td><a href="./defs.md#defs-tooltip">tooltipTextFunction</a></td><td>Callback function that handles the dynamic text appearing in the tooltip.</td>
     </tr>
     <tr>
       <td>options</td>
@@ -38,16 +38,17 @@
 
 &nbsp; &nbsp; &nbsp; &nbsp;**Returns:** Self, for chaining functions.
 
-<a name="api-sg-addTooltipToLines"></a>
-**`SimpleGraph.addTooltipToLines(tooltipFunction[, options])`** : Add tooltip function to the lines on the graph.
+<a name="addTooltipToLines">#</a> *sg*.**addTooltipToLines**(*tooltipFunction*[, *options*])
 
-<table>
+Add tooltip function to the lines on the graph.
+
+<table style="font-size:0.9em;">
   <tbody>
     <tr>
       <th>Name</th><th>Type</th><th>Description</th>
     </tr>
     <tr>
-      <td>tooltipFunction</td><td><a href="./defs.md#api-sg-defs-tooltip">tooltipTextFunction</a></td><td>Callback function that handles the dynamic text appearing in the tooltip.</td>
+      <td>tooltipFunction</td><td><a href="./defs.md#defs-tooltip">tooltipTextFunction</a></td><td>Callback function that handles the dynamic text appearing in the tooltip.</td>
     </tr>
     <tr>
       <td>options</td>
@@ -75,16 +76,17 @@
 
 &nbsp; &nbsp; &nbsp; &nbsp;**Returns:** Self, for chaining functions.
 
-<a name="api-sg-addTooltipToAreas"></a>
-**`SimpleGraph.addTooltipToAreas(tooltipFunction[, options])`** : Add tooltip function to the areas on the graph.
+<a name="addTooltipToAreas">#</a> *sg*.**addTooltipToAreas**(*tooltipFunction*[, *options*])
 
-<table>
+Add tooltip function to the areas on the graph.
+
+<table style="font-size:0.9em;">
   <tbody>
     <tr>
       <th>Name</th><th>Type</th><th>Description</th>
     </tr>
     <tr>
-      <td>tooltipFunction</td><td><a href="./defs.md#api-sg-defs-tooltip">tooltipTextFunction</a></td><td>Callback function that handles the dynamic text appearing in the tooltip.</td>
+      <td>tooltipFunction</td><td><a href="./defs.md#defs-tooltip">tooltipTextFunction</a></td><td>Callback function that handles the dynamic text appearing in the tooltip.</td>
     </tr>
     <tr>
       <td>options</td>
@@ -111,13 +113,14 @@
 
 &nbsp; &nbsp; &nbsp; &nbsp;**Returns:** Self, for chaining functions.
 
-<a name="api-sg-functions-highlight"></a>
+
 ## Highlight Functions ##
 
-<a name="api-sg-highlightPoints"></a>
-**`SimpleGraph.highlightPoints(series[, validationCallback[, size[, fill[, stylesDict]]]])`** : Highlights points by drawing new SVG over points. Highlights a given data series, using optional `validationCallback` to filter points within data series. Note that while `size`, `fill`, and `stylesDict` are all optional, if none are supplied, the highlight will look exactly like the point and no difference will be noticeable.
+<a name="highlightPoints">#</a> *sg*.**highlightPoints**(*series*[, *validationCallback*[, *size*[, *fill*[, *stylesDict*]]]])
 
-<table>
+Highlights points by drawing new SVG over points. Highlights a given data series, using optional `validationCallback` to filter points within data series. Note that while `size`, `fill`, and `stylesDict` are all optional, if none are supplied, the highlight will look exactly like the point and no difference will be noticeable.
+
+<table style="font-size:0.9em;">
   <tbody>
     <tr>
       <th>Name</th><th>Type</th><th>Description</th>
@@ -126,10 +129,10 @@
       <td>series</td><td>string</td><td>The name of the data series.</td>
     </tr>
     <tr>
-      <td>validationCallback</td><td>function</td><td>The callback function to validate whether to highlight given point. Passed argument of <a href="./defs.md#api-sg-defs-pointdata">point data</a> for given point. Return true to include point. If callback is null, assumes all points to be included.</td>
+      <td>validationCallback</td><td>function</td><td>The callback function to validate whether to highlight given point. Passed argument of <a href="./defs.md#defs-pointdata">point data</a> for given point. Return true to include point. If callback is null, assumes all points to be included.</td>
     </tr>
     <tr>
-      <td>size</td><td>number|function</td><td>Point size or callback function to determine point size. Called in scope such that `this` will refer to the <a href="./defs.md#api-sg-defs-pointdata">point data</a>. If null, uses assigned `pointsize` for point data.</td>
+      <td>size</td><td>number|function</td><td>Point size or callback function to determine point size. Called in scope such that `this` will refer to the <a href="./defs.md#defs-pointdata">point data</a>. If null, uses assigned `pointsize` for point data.</td>
     </tr>
     <tr>
       <td>fill</td><td>string</td><td>Fill value. If null, uses same as for data series.</td>
@@ -142,23 +145,14 @@
 
 &nbsp; &nbsp; &nbsp; &nbsp;**Returns:** Self, for chaining functions.
 
-<a name="api-sg-removeHighlightPoints"></a>
-**`SimpleGraph.removeHighlightPoints()`** : Remove any highlights on points.
+<a name="removeHighlightPoints">#</a> *sg*.**removeHighlightPoints**()
+
+Remove any highlights on points.
 
 &nbsp; &nbsp; &nbsp; &nbsp;**Returns:** Self, for chaining functions.
 
-<a name="api-sg-removeHighlights"></a>
-**`SimpleGraph.removeHighlights()`** : Remove all highlights.
+<a name="removeHighlights">#</a> *sg*.**removeHighlights**()
+
+Remove all highlights.
 
 &nbsp; &nbsp; &nbsp; &nbsp;**Returns:** Self, for chaining functions.
-
-
-
-<style>
-  table { border-collapse: collapse; font-size: 0.9em; margin-left: 2.5em; }
-  th { text-align: left; background-color: #bbb; }
-  td h5 { margin: 0.4em 0.2em; font-size: 1.05em; font-style: italic; }
-  td table { font-size: 1em; margin: 0.2em; width: 100%; }
-  td table th { background-color: #d6d6d6; border-color: #ddd; }
-  td table td { background-color: #fff; border-color: #ddd; }
-</style>
